backend/requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Utilities
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
PyJWT>=2.8,<3.0
pytz>=2024.1

# Django dependencies (usually auto-installed)
asgiref>=3.7,<4.0
sqlparse>=0.5,<1.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
responses>=0.25,<1.0
